+++ packages/cli-plugin-deploy/src/index.ts
@@ -47,9 +47,11 @@ export class DeployPlugin extends BasePlugin {
     if (domain === 'auto') {
       return;
     }
-    console.log('\nDomain bind: ' + domain);
+    this.core.cli.log('\nDomain bind: ' + domain);
     for (const path of allPaths) {
-      console.log(`- path bind: http://${domain}/${path.replace(/^\//, '')}`);
+      this.core.cli.log(
+        `- path bind: http://${domain}/${path.replace(/^\//, '')}`
+      );
     }
   }
 }